convert anon recruitment modal to section
